% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{get_hdi}
\alias{get_hdi}
\title{get_hdi
Computes HDI for vector vec and hdi_level (e.g. 0.95). Taken (and renamed)
from "Doing Bayesian Analysis", section 25.2.3 R code for computing HDI of
a MCMC sample}
\usage{
get_hdi(vec, hdi_level)
}
\arguments{
\item{vec}{numeric vector}

\item{hdi_level}{number, highest density interval (HDI) (default = 0.95)}
}
\value{
vector with two elements (low, high interval of HDI)
}
\description{
get_hdi
Computes HDI for vector vec and hdi_level (e.g. 0.95). Taken (and renamed)
from "Doing Bayesian Analysis", section 25.2.3 R code for computing HDI of
a MCMC sample
}
