% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{get_multinomial_rng}
\alias{get_multinomial_rng}
\title{get_multinomial_rng}
\usage{
get_multinomial_rng(af, B, N)
}
\arguments{
\item{af}{vector, allele frequencies, sum(af)=1}

\item{B}{integer, number of simulated samples}

\item{N}{integer, number of alleles in simulated sample}
}
\value{
y_rng numeric matrix, number of alleles in each simulated sample
}
\description{
This function simulates B samples (rows of counts of matrix y) from
a multivariate distribution. Each count belongs to one of K categories
(K columns of matrix y). The data is simulated based on the given
simplex af and N number of total alleles.
}
\examples{
get_multinomial_rng(K=10, af=rep(0.1, times = 10), B=100, N=1000)
}
