% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{get_power}
\alias{get_power}
\title{get_power}
\usage{
get_power(y, y_rng, a, rng_draws = 10^4)
}
\arguments{
\item{y, }{numeric vector, Multinomially distributed sample (background)}

\item{y_rng, }{numeric matrix, Multinomially distributed samples (rows)}

\item{a, }{numeric vector, Dirichlet distribution parameters (alpha)}

\item{rng_draws, }{integer, number of posterior draws (default 10,000 )}
}
\value{
data.frame of posterior summaries
}
\description{
get_power
}
