% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R, R/power_loop.R
\name{get_power_analysis}
\alias{get_power_analysis}
\title{get_power_analysis}
\usage{
get_power_analysis(y, y_rng, a, rng_draws = 10^4, n)

get_power_analysis(y, y_rng, a, rng_draws = 10^4, n)
}
\arguments{
\item{y, }{numeric vector, Multinomially distributed sample (background)}

\item{y_rng, }{numeric matrix, Multinomially distributed samples (rows)}

\item{a, }{numeric vector, Dirichlet distribution parameters (alpha)}

\item{rng_draws, }{integer, number of posterior draws (default 10,000)}

\item{n, }{integer, study size to investigate}
}
\value{
data.frame of posterior summaries

data.frame of posterior summaries
}
\description{
get_power_analysis

get_power_analysis
}
