% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{get_power_analysis}
\alias{get_power_analysis}
\title{get_power_analysis}
\usage{
get_power_analysis(
  ns,
  theta,
  gamma,
  alpha,
  B,
  rng_draws = 10^4,
  cores = 1,
  verbose = T
)
}
\arguments{
\item{ns, }{integer vector, total number of alleles in simulated samples}

\item{theta, }{vector of reals, allele frequencies of the simulated samples}

\item{B, }{integer, number of simulations to use}

\item{rng_draws, }{integer, number of posterior draws (default 10,000)}

\item{cores, }{integer, number of cores for multicore execution}

\item{y, }{vector of integers, allele counts of the background sample}

\item{a, }{vector of reals, Dirichlet distribution parameters (alpha)}
}
\description{
get_power_analysis
}
