% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{get_power_summary}
\alias{get_power_summary}
\title{get_power_summary}
\usage{
get_power_summary(p)
}
\arguments{
\item{p, }{object returned by function get_power_analysis}
}
\value{
data.frame
}
\description{
get_power_summary
}
\examples{
get_power_summary(p)
}
